update
add session
add redirection when session inactive

+++ page/admin_produk_delete.php
@@ -1,5 +1,8 @@
 <?php
 include('../script/produk.php');
+include('../script/session.php');
+
+check_admin_session();
 
 if(delete_data_produk($_GET['id'])) {
     header('location: ./admin_produk_data.php');
