+++ page/admin_promo_delete.php
@@ -1,5 +1,8 @@
 <?php
 include('../script/promo.php');
+include('../script/session.php');
+
+check_admin_session();
 
 if(delete_data_promo($_GET['id'])) {
     header('location: ./admin_promo_data.php');
